backend/requirements.txt
Flask==3.0.3
python-docx==1.1.2
Flask-Cors==4.0.1
pymongo==4.7.3
dnspython==2.6.1
python-dotenv==1.0.1
Flask-PyMongo==2.3.0
Werkzeug==3.0.3
Flask-JWT-Extended==4.6.0
pyotp==2.9.0